requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
pandas==2.0.3
numpy==1.25.2
matplotlib==3.7.1
scikit-learn==1.5.0
seaborn==0.13.1
Pillow==9.4.0
Jinja2
tqdm==4.66.4
tabulate==0.9.0
imgaug==0.4.0
albumentations==1.4.8
ultralytics==8.2.32
comet-ml==3.43.1
wget==3.2

# Only for local notebook, not for azure function
# ipywidgets